Removing the mvp files and refactoring code
