+++ app/src/main/java/com/mn2square/videolistingmvp/mvvm/ViewMvpSearch.java
@@ -1,4 +1,4 @@
-package com.mn2square.videolistingmvp.activity.views;
+package com.mn2square.videolistingmvp.mvvm;
 
 import android.view.MenuItem;
