+++ app/src/main/java/com/mn2square/videolistingmvp/mvvm/pojo/VideoListInfo.java
@@ -1,4 +1,4 @@
-package com.mn2square.videolistingmvp.activity.presenter.manager.pojo;
+package com.mn2square.videolistingmvp.mvvm.pojo;
 
 import java.util.ArrayList;
 import java.util.HashMap;
