+++ app/src/main/java/com/mn2square/videolistingmvp/mvvm/swipetabfragments/folderlistfragment/FolderListFragmentViewModel.java
@@ -1,33 +1,20 @@
 package com.mn2square.videolistingmvp.mvvm.swipetabfragments.folderlistfragment;
 
-import com.mn2square.videolistingmvp.R;
-import com.mn2square.videolistingmvp.activity.presenter.manager.pojo.VideoListInfo;
+import com.mn2square.videolistingmvp.mvvm.pojo.VideoListInfo;
 import com.mn2square.videolistingmvp.mvvm.repository.VideoListRepository.VideoListRepositoryImpl;
-import com.mn2square.videolistingmvp.swipetabfragments.adapters.FolderListAdapter;
-import com.mn2square.videolistingmvp.swipetabfragments.folderlistfragment.ObservableFolderList.ObservableExpandableListView;
-import com.mn2square.videolistingmvp.swipetabfragments.folderlistfragment.views.FolderListFragmentView;
 
 import android.app.Application;
 import android.arch.lifecycle.AndroidViewModel;
 import android.arch.lifecycle.MediatorLiveData;
-import android.arch.lifecycle.MutableLiveData;
 import android.arch.lifecycle.Observer;
-import android.content.Context;
 import android.content.SharedPreferences;
-import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.util.Log;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
-import java.util.HashMap;
-import java.util.List;
 
 /**
  * Created by nitinagarwal on 3/15/17.
