--- app/src/main/java/com/mn2square/videolistingmvp/swipetabfragments/VideoListFragmentInterface/VideoListFragmentInterface.java
@@ -1,11 +1,0 @@
-package com.mn2square.videolistingmvp.swipetabfragments.VideoListFragmentInterface;
-
-import com.mn2square.videolistingmvp.activity.presenter.manager.pojo.VideoListInfo;
-
-/**
- * Created by nitinagarwal on 4/9/17.
- */
-
-public interface VideoListFragmentInterface {
-    void bindVideoList(VideoListInfo videoListInfo);
-}
