--- app/src/main/java/com/mn2square/videolistingmvp/viewmvp/ViewMvp.java
@@ -1,13 +1,0 @@
-package com.mn2square.videolistingmvp.viewmvp;
-
-import android.os.Bundle;
-import android.view.View;
-
-/**
- * Created by nitinagarwal on 4/9/17.
- */
-
-public interface ViewMvp {
-    View getRootView();
-    Bundle getViewState();
-}
